m2vcut-gui: constant font size whatever the [Xft.]dpi is
The font "Monospace bold 12" was fine when display "dpi" setting was
~96 (to me this looks like long-time default for lo-dpi displays).

Whenever dpi setting was higher (enough) pango scaled the font
in question bigger, resulting ugly layout.

Now the font is absolutely sized -- it may look small in very high
dpi displays, but the layout is consistent now.

Thanks to Juhani Mäkelä for reporting the problem and basically
showing the code position where to fix it.